% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wordleUtilities.R
\name{getGuess}
\alias{getGuess}
\title{Get a guess (word) from a bot}
\usage{
getGuess(bot, wordleEvaluatorResponse, validWordVector, maxBadWords = 100)
}
\arguments{
\item{bot}{Name of bot}

\item{wordleEvaluatorResponse}{list representing the response from the wordleEvaluator function.}

\item{validWordVector}{vector of valid words}

\item{maxBadWords}{Integer representing maximum illegal words which may be attempted}
}
\value{
string which is the guess of a bot
}
\description{
Get a guess (word) from a bot
}
