% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wordleUtilities.R
\name{testWord}
\alias{testWord}
\title{Test a guess}
\usage{
testWord(guess, wordleWord, n = 5)
}
\arguments{
\item{guess}{Word suggested by bot (a guess)}

\item{wordleWord}{Word to be found (the correct Wordle answer)}

\item{n}{length of a word (default = 5)}
}
\value{
five character string
}
\description{
A five character string will be returned consisting of 0s, 1s and 2s.  A 0 represents a letter not in the word.
A 1 represents a letter in the word but in the wrong location and a 2 is a letter in the correct location. So
a correct guess would return "22222".  A guess of "ABCDE" when the Wordle is "VWXYZ" would be "00000" while
"ABCDE" guessed for "EDCBA" would return "11211"
}
