% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wordleUtilities.R
\name{wordleChallenge}
\alias{wordleChallenge}
\title{Wordle Challenge}
\usage{
wordleChallenge(botList, challengeWords, wordList)
}
\arguments{
\item{botList}{list of bots}

\item{challengeWords}{A character vector of five letter words each bot is to try to guess.}

\item{wordList}{A character vector of valid words.}
}
\value{
List with three elements.  'averageScores' is a data frame with a row for each bot and columns for the name of the bot,
the AverageScore of the bot, and the Rank of the bot.  'scores' is a data frame with a row for each challengeWords and columns for
the challenge word for each bot's score for that word.
}
\description{
This is the main function of the package.  It takes a list of bots and a list (vector) of five letter words and generates an
average score for each bot. Essentially it repeatedly calls the wordleEvaluator function for each bot for each word.
}
\examples{
\dontrun{wordleChallenge(botList, challengeWords, wordList)}
}
