% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wordleUtilities.R
\name{wordleEvaluator}
\alias{wordleEvaluator}
\title{Wordle Evaluator}
\usage{
wordleEvaluator(
  bot,
  wordleWord,
  validWordVector,
  maxBadWords = 100,
  failScore = 10
)
}
\arguments{
\item{bot}{Name of function to use to solve wordle puzzles}

\item{wordleWord}{The world which is the solution to the puzzle}

\item{validWordVector}{A vector of valid 5 letter words}

\item{maxBadWords}{Integer representing maximum illegal words which may be attempted}

\item{failScore}{Score if the bot cannot solve the puzzle.}
}
\value{
List with 3 elements:  response, guesses, score
}
\description{
This is a core function of the package.  It takes the name of a bot and a single five letter "wordleWord" and generates a score
equal to the number of tries is takes to guess the word, or the failScore (default = 10) if the puzzle is not solved.
If a guess is not in the validWordVector then it is a "bad word".  If the bot guesses maxBadWords that is considered a failure.
}
\details{
To retrieve a guess from the bot function, wordleEvaluator calls the bot function with two parameters (so the bot must
accept 2 parameters).  The first parameter is a list.  The second is a validWordVector.  The list will have three elements matching
the output of the wordleEvaluator function.

wordleEvaluator returns a list with 3 elements.  The first element is 'response' a character vector. Initially this is NULL.  After each guess by
the bot, a string will be added to the response vector.  The string will be five characters consisting of 0s, 1s and 2s.
A 0 represents a letter not in the word. A 1 represents a letter in the word but in the wrong location and a 2 is a letter in the correct location. So
a correct guess would return "22222".  A guess of "ABCDE" when the WordleWord is "VWXYZ" would be "00000" while
"ABCDE" guessed for "EDCBA" would return "11211".

The 'guesses' element in the list that is returned is a character vector of the guesses provided to the wordleEvaluator function by the bot. Initially
guesses will be NULL.

The 'score' element is an integer from 1 to 6 if the bot successfully solved the puzzle.  The value represents which try was successful. The
score will be the failScore (default = 10) is the bot was unsuccessful.

Hint/Suggestion:  The bot function may want to test either the response or the guesses elements for being NULL. If NULL, then the bot should return its
first guess.

The concept of this package is that this function will be called for many (e.g. 1000) words in the validWordVector for each bot in a challenge. The
bot with the lowest average wins.
}
\examples{
\dontrun{wordleEvaluator(exampleBot, "CYNIC", validWordVector)}
}
