% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{wordList}
\alias{wordList}
\title{Word list with over 15,000 five letter words.}
\format{
A character vector with 15,918 elements
}
\source{
somewhere on the internet
}
\usage{
data(wordList)
}
\description{
A dataset containing over 15,000 five letter words.  Useful for Wordle-type games and problems.
}
\keyword{datasets}
